build(ci): checkout before installing browser-tools
For some reason the checkout step all of a sudden started giving: "Directory
(/home/circleci/project) you are trying to checkout to is not empty and not a git repository".
After a re-run with SSH I noticed there's a LICENSE.chromedriver file in the project directory
which led me to this:
https://discuss.circleci.com/t/git-checkout-during-build-suddenly-no-longer-working/46680.
That suggests moving the checkout step up above browser-tools installation.