fix: change streamPhases type to array of tuples, use humanStringify
Config streamPhases type change to enable validation of time values as numbers.
And to facilitate working with it down the line in TypeScript.
Use humanStringify() to support showing values like Infinity for getPlayback duration.
Fix LiveryBridgeInteractive to enable selecting quality -1 for ABR.
Have MockPlayerBridge return an error as expected when selecting an illegal quality index.
And tweak some other things for testing purposes.